refactor and cleanup config loading and tests
