+++ triggers/service/src/main/scala/com/digitalasset/daml/lf/engine/trigger/TriggerServiceAppConf.scala
@@ -1,0 +1,137 @@
+// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.lf.engine.trigger
+
+import akka.http.scaladsl.model.Uri
+import com.daml.dbutils.JdbcConfig
+import com.daml.lf.speedy.Compiler
+import com.daml.platform.services.time.TimeProviderType
+import pureconfig.{ConfigReader, ConvertHelpers}
+import com.daml.auth.middleware.api.{Client => AuthClient}
+import pureconfig.error.FailureReason
+import pureconfig.generic.semiauto.deriveReader
+
+import java.nio.file.Path
+import java.time.Duration
+import scala.concurrent.duration.FiniteDuration
+
+private[trigger] object LedgerApiConfig {
+  lazy implicit val ledgerApiCfgReader: ConfigReader[LedgerApiConfig] =
+    deriveReader[LedgerApiConfig]
+}
+private[trigger] final case class LedgerApiConfig(address: String, port: Int)
+
+private[trigger] object AuthorizationConfig {
+  final case object AuthConfigFailure extends FailureReason {
+    val description =
+      "You must specify either just auth-common-uri or both auth-internal-uri and auth-external-uri"
+  }
+
+  def isValid(ac: AuthorizationConfig): Boolean = {
+    (ac.authCommonUri.isDefined && ac.authExternalUri.isEmpty && ac.authInternalUri.isEmpty) ||
+    (ac.authCommonUri.isEmpty && ac.authExternalUri.nonEmpty && ac.authInternalUri.nonEmpty)
+  }
+
+  lazy implicit val uriCfgReader: ConfigReader[Uri] =
+    ConfigReader.fromString[Uri](ConvertHelpers.catchReadError(s => Uri(s)))
+
+  lazy implicit val redirectToLoginCfgReader: ConfigReader[AuthClient.RedirectToLogin] =
+    ConfigReader.fromString[AuthClient.RedirectToLogin](
+      ConvertHelpers.catchReadError(s => Cli.redirectToLogin(s))
+    )
+
+  lazy implicit val authCfgReader: ConfigReader[AuthorizationConfig] =
+    deriveReader[AuthorizationConfig].emap { ac =>
+      Either.cond(isValid(ac), ac, AuthConfigFailure)
+    }
+}
+private[trigger] final case class AuthorizationConfig(
+    authInternalUri: Option[Uri] = None,
+    authExternalUri: Option[Uri] = None,
+    authCommonUri: Option[Uri] = None,
+    authRedirect: AuthClient.RedirectToLogin = AuthClient.RedirectToLogin.No,
+    authCallbackUri: Option[Uri] = None,
+    maxPendingAuthorizations: Int = Cli.DefaultMaxAuthCallbacks,
+    authCallbackTimeout: FiniteDuration = Cli.DefaultAuthCallbackTimeout,
+)
+
+private[trigger] object TriggerServiceAppConf {
+  lazy implicit val compilerCfgReader: ConfigReader[Compiler.Config] =
+    ConfigReader.fromString[Compiler.Config](ConvertHelpers.catchReadError { s =>
+      s.toLowerCase() match {
+        case "default" => Compiler.Config.Default
+        case "dev" => Compiler.Config.Dev
+        case s =>
+          throw new IllegalArgumentException(
+            s"Value '$s' for compiler-config is not one of 'default' or 'dev'"
+          )
+      }
+    })
+  lazy implicit val timeProviderTypeCfgReader: ConfigReader[TimeProviderType] =
+    ConfigReader.fromString[TimeProviderType](ConvertHelpers.catchReadError { s =>
+      s.toLowerCase() match {
+        case "static" => TimeProviderType.Static
+        case "wall-clock" => TimeProviderType.WallClock
+        case s =>
+          throw new IllegalArgumentException(
+            s"Value '$s' for time-provider-type is not one of 'static' or 'wall-clock'"
+          )
+      }
+    })
+  lazy implicit val jdbcCfgReader: ConfigReader[JdbcConfig] = deriveReader[JdbcConfig]
+  lazy implicit val serviceCfgReader: ConfigReader[TriggerServiceAppConf] =
+    deriveReader[TriggerServiceAppConf]
+}
+
+/* An intermediate config representation allowing us to define our HOCON config in a more modular fashion,
+  this eventually gets mapped to `ServiceConfig`
+ */
+private[trigger] final case class TriggerServiceAppConf(
+    darPaths: List[Path],
+    address: String,
+    port: Int,
+    portFile: Option[Path] = None,
+    ledgerApi: LedgerApiConfig,
+    authorization: AuthorizationConfig = AuthorizationConfig(),
+    maxInboundMessageSize: Int = Cli.DefaultMaxInboundMessageSize,
+    minRestartInterval: FiniteDuration = Cli.DefaultMinRestartInterval,
+    maxRestartInterval: FiniteDuration = Cli.DefaultMaxRestartInterval,
+    maxHttpEntityUploadSize: Long = Cli.DefaultMaxHttpEntityUploadSize,
+    httpEntityUploadTimeout: FiniteDuration = Cli.DefaultHttpEntityUploadTimeout,
+    timeProviderType: TimeProviderType = TimeProviderType.WallClock,
+    ttl: FiniteDuration = Cli.DefaultCommandTtl,
+    initDb: Boolean = false,
+    triggerStore: Option[JdbcConfig] = None,
+    allowExistingSchema: Boolean = false,
+    compilerConfig: Compiler.Config = Compiler.Config.Default,
+) {
+  def toServiceConfig: ServiceConfig = {
+    ServiceConfig(
+      darPaths,
+      address,
+      port,
+      ledgerApi.address,
+      ledgerApi.port,
+      authorization.authInternalUri,
+      authorization.authExternalUri,
+      authorization.authCommonUri,
+      authorization.authRedirect,
+      authorization.authCallbackUri,
+      maxInboundMessageSize,
+      minRestartInterval,
+      maxRestartInterval,
+      authorization.maxPendingAuthorizations,
+      authorization.authCallbackTimeout,
+      maxHttpEntityUploadSize,
+      httpEntityUploadTimeout,
+      timeProviderType,
+      Duration.ofSeconds(ttl.toSeconds),
+      initDb,
+      triggerStore,
+      portFile,
+      allowExistingSchema,
+      compilerConfig,
+    )
+  }
+}
