+++ triggers/service/src/test-suite/resources/trigger-service.conf
@@ -62,6 +62,9 @@
     //Sets both the internal and external auth URIs.
     //auth-common-uri = "https://oauth2/common-uri"
 
+    // Auth Client to redirect to login , defaults to no
+    auth-redirect = "yes"
+
     //Sets the internal auth URIs (used by the trigger service to connect directly to the middleware). Overrides value set by auth-common
     auth-internal-uri = "https://oauth2/internal-uri"
     //Sets the external auth URI (the one returned to the browser). overrides value set by auth-common.
